Move solve.py into physics/
- The simulation (event_based/) should not know about details like the shape of the
  collision surface
